feat(logging): implement file logging
